src/cloud_function/requirements.txt
protobuf
torch
transformers
google-cloud-storage